test(tracing): isolate tool execution hierarchy coverage
